AI/requirements.txt
tflite-runtime
fastapi
uvicorn
Pillow
python-multipart
numpy
opencv-python-headless